requirements.txt


pandas
numpy
matplotlib
PyYAML
statsmodels
scikit-learn
